ai_hackathon-main/ai_hackathon-main/requirements.txt
fastapi
uvicorn
pdfplumber
pymupdf
camelot-py[cv]
pandas
pillow
pdf2image
easyocr
opencv-python
python-multipart
google-genai
python-dotenv
